rename filereader.cc[h] to logreader.cc[h] (#22279)
